Add files via upload
This document is part of the ISO/IEC 27001:2022 Lead Implementer Course. It focuses on understanding and applying the principles of information security management within a specified context. The content involves identifying relevant interested parties and their needs, analyzing risk factors, and implementing appropriate security measures in line with ISO 27001:2022 standards. It is designed to help participants enhance their practical skills in auditing and managing information security in an organizational setting.