Add files via upload
This repository contains materials and resources for Exercise 6 of the ISO 27001:2022 Lead Implementer Course. It includes a detailed walkthrough of the exercise, focusing on the identification and management of information security risks within an organization. All relevant documents, templates, and guidelines are provided to assist in understanding and completing this exercise